added three testing credentials

+++ Controller/User/userCont.js
@@ -279,6 +279,21 @@ exports.loginByMobile = async (req, res) => {
       });
     }
 
+    // Testing
+    if (
+      mobileNumber === "9675355345" || // Ankush
+      mobileNumber === "8171156708" || // Laxmi
+      mobileNumber === "8938065100" // Amit
+    ) {
+      return res.status(200).send({
+        success: true,
+        message: `OTP send successfully! Valid for ${
+          OTP_VALIDITY_IN_MILLISECONDS / (60 * 1000)
+        } minutes!`,
+        data: { mobileNumber: mobileNumber },
+      });
+    }
+
     // Generate OTP for Email
     const otp = generateFixedLengthRandomNumber(OTP_DIGITS_LENGTH);
     // Sending OTP to mobile number
@@ -314,32 +329,7 @@ exports.verifyMobileOTP = async (req, res) => {
       return res.status(400).send(error.details[0].message);
     }
     const { mobileNumber, otp } = req.body;
-    // Only for testing
-    if (mobileNumber === "1133557799") {
-      const token = createAccessToken(
-        "user",
-        "playstoretester@gmail.com",
-        "67165fa7b1474b741db09d0a"
-      );
-      const refreshToken = createRefreshToken(
-        "user",
-        "67165fa7b1474b741db09d0a"
-      );
-      return res.status(200).json({
-        success: true,
-        AccessToken: token,
-        refreshToken,
-        user: {
-          _id: "67165fa7b1474b741db09d0a",
-          name: "Tester For Play Store",
-          email: "playstoretester@gmail.com",
-          isLicenseVerified: false,
-          mobileNumber: "1133557799",
-          role: undefined,
-        },
-        message: `Welcome, Tester For Play Store`,
-      });
-    }
+
     // Is Email Otp exist
     const isOtp = await OTP.findOne({
       otp: otp,
@@ -363,16 +353,25 @@ exports.verifyMobileOTP = async (req, res) => {
         message: "No Details Found. Register Now!",
       });
     }
-    // is email otp expired?
-    const isOtpExpired = new Date().getTime() > parseInt(isOtp.validTill);
-    if (isOtpExpired) {
+
+    if (
+      mobileNumber === "9675355345" || // Ankush
+      mobileNumber === "8171156708" || // Laxmi
+      mobileNumber === "8938065100" // Amit
+    ) {
+      // Do Nothing
+    } else {
+      // is email otp expired?
+      const isOtpExpired = new Date().getTime() > parseInt(isOtp.validTill);
+      if (isOtpExpired) {
+        await OTP.deleteMany({ receiverId: isOtp.receiverId });
+        return res.status(400).send({
+          success: false,
+          message: `OTP expired!`,
+        });
+      }
       await OTP.deleteMany({ receiverId: isOtp.receiverId });
-      return res.status(400).send({
-        success: false,
-        message: `OTP expired!`,
-      });
     }
-    await OTP.deleteMany({ receiverId: isOtp.receiverId });
 
     const refreshToken = createRefreshToken("user", user._id);
     // Update user
